requirements.txt
fastapi>=0.110.0
uvicorn>=0.28.0
pydantic>=2.6.0
streamlit>=1.32.0
pandas>=2.2.0
numpy>=1.26.0
requests>=2.31.0
scikit-learn>=1.4.0
pyyaml>=6.0
